+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- `ConditionLayoutCategory` component. 
 
 ## [2.3.0] - 2021-08-02
 ### Added
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `ConditionLayoutBinding` component. 
 
+ 
 ## [2.1.3] - 2021-01-27
 ### Fixed
 - Avoid breaking if product context is not filled yet.